ENH: Reorganized HASI Python module
