+++ src/hasi/hasi/__init__.py
@@ -1,0 +1,3 @@
+"""High-throughput Applications for Skeletal Imaging (HASI) Python content"""
+
+__version__ = '0.1'
