+++ src/hasi/hasi/diffeoregistrar.py
